Restructuring and clean-ups
